\name{log_enable}
\alias{log_enable}
\title{Enable logging at a specific level}
\usage{
log_enable(level = "INFO")
}
\arguments{
\item{level}{Character scalar. One of "TRACE", "DEBUG", "INFO", "WARN", "ERROR".}
}
\description{
Sets the package-wide logging verbosity.
}
\value{
No return value, called for side effects
}