% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_enable}
\alias{log_enable}
\title{Enable logging at a specific level}
\usage{
log_enable(level = "INFO")
}
\arguments{
\item{level}{Character scalar. One of "TRACE", "DEBUG", "INFO", "WARN", "ERROR".}
}
\value{
(Invisibly) the level set.
}
\description{
Sets the package-wide logging verbosity.
}
